backend/pyproject.toml
[tool.poetry]
name = "contractlens-backend"
version = "0.1.0"
description = "AI-powered contract review and risk analysis API"
authors = ["Vijay Prabhu <vijayprabhumarikkani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
supabase = "^2.3.0"
openai = "^1.10.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
python-multipart = "^0.0.6"
python-docx = "^1.1.0"
pymupdf = "^1.23.0"
redis = "^5.0.0"
httpx = "^0.26.0"
tenacity = "^8.2.0"
pgvector = "^0.2.4"
psycopg = {extras = ["binary"], version = "^3.3.2"}
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
